Update README.md
The primary objective of this project is to develop predictive models that accurately classify mushrooms as edible or poisonous based on their features. My aim is to utilize various machine learning algorithms, including hyperparameter tuning, to improve model performance and ensure robust evaluation. Additionally, the project seeks to identify key features that contribute to the classification, offering insights into mushroom characteristics and safety